update: create outbond strategy

+++ src/middlewares/multer.js
@@ -4,6 +4,9 @@ const fs = require("fs");
 const imageHandler = multer({
   storage: multer.diskStorage({
     destination: (req, file, cb) => {
+      if (!fs.existsSync(`${__dirname}/../../public/images`)) {
+        fs.mkdirSync(`${__dirname}/../../public/images`, { recursive: true });
+      }
       cb(null, `${__dirname}/../../public/images`);
     },
     filename: (req, file, cb) => {
@@ -21,11 +24,15 @@ const imageHandler = multer({
 });
 
 const deleteFile = (filePath) => {
+  if (!fs.existsSync(filePath))
+    return "Image File not found, but data was deleted!";
   fs.unlink(filePath, (err) => {
     if (err) {
       console.log(err);
+      return `Internal Error when deleting file: ${err}`;
     } else {
       console.log("File deleted");
+      return "Image File deleted succesfully";
     }
   });
 };
